[tests] Update this test to be more correct
We cannot have proper piece picking in metadata mode as we do
not acutally know how many pieces the torrent has. This test
had begun to fail as other parts of the codebase handle this
case more correctly (enforcing the invariant). As such it's
time to update this test so it uses a proper torrent.

+++ src/MonoTorrent.Tests/Client/TestRig.cs
@@ -547,12 +547,12 @@ namespace MonoTorrent.Client
             return new TorrentManager (editor.ToTorrent (), "", new TorrentSettings ());
         }
 
-        internal static TestRig CreateSingleFile (int torrentSize, int pieceLength)
+        internal static TestRig CreateSingleFile (long torrentSize, int pieceLength)
         {
             return CreateSingleFile (torrentSize, pieceLength, false);
         }
 
-        internal static TestRig CreateSingleFile (int torrentSize, int pieceLength, bool metadataMode)
+        internal static TestRig CreateSingleFile (long torrentSize, int pieceLength, bool metadataMode)
         {
             TorrentFile[] files = StandardSingleFile ();
             files[0] = new TorrentFile (files[0].Path, torrentSize);
@@ -564,7 +564,7 @@ namespace MonoTorrent.Client
             return new TestRig ("", pieceLength, StandardWriter (), StandardTrackers (), StandardMultiFile (), metadataMode);
         }
 
-        internal static TorrentManager CreateSingleFileManager (int torrentSize, int pieceLength)
+        internal static TorrentManager CreateSingleFileManager (long torrentSize, int pieceLength)
         {
             return CreateSingleFile (torrentSize, pieceLength, false).Manager;
         }
